+++ Altinn.Platform.Events.IsolatedFunctions/Altinn.Platform.Events.IsolatedFunctions.csproj
