+++ Altinn.Platform.Events.IsolatedFunctions/EventsOutbound.cs
