+++ Altinn.Platform.Events.IsolatedFunctions/Properties/launchSettings.json
