+++ Altinn.Platform.Events.IsolatedFunctions/Properties/serviceDependencies.json
