+++ Altinn.Platform.Events.IsolatedFunctions/Properties/serviceDependencies.local.json
