+++ Altinn.Platform.Events.IsolatedFunctions/host.json
