+++ Altinn.Platform.Events.IsolatedFunctions/local.settings.json
