+++ src/Events.Functions/Altinn.Platform.Events.Functions.csproj
@@ -1,13 +1,13 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
     <TargetFramework>net9.0</TargetFramework>
-    <AzureFunctionsVersion>v4</AzureFunctionsVersion>
     <GenerateDocumentationFile>true</GenerateDocumentationFile>
     <ProjectGuid>{1F3A1E44-1333-46B8-8A70-08BE01C074F6}</ProjectGuid>
   </PropertyGroup>
   <ItemGroup>
     <!-- Upgrading to 1.1.5 will cause the function to fail, must upgrade function to .net 7 first-->
     <PackageReference Include="Altinn.Common.AccessTokenClient" Version="[1.1.3]" />
+    <PackageReference Include="Azure.Identity" Version="1.14.1" />
 	<!-- Upgrading to 4.8.0 will cause the function to fail, must upgrade function from .Net 6 first-->
     <PackageReference Include="Azure.Security.KeyVault.Secrets" Version="4.8.0" />
 	<!-- Upgrading to 4.8.0 will cause the function to fail, must upgrade function from .Net 6 first-->
@@ -15,9 +15,6 @@
     <PackageReference Include="CloudNative.CloudEvents.SystemTextJson" Version="2.8.0" />
     <PackageReference Include="Microsoft.Extensions.Http" Version="9.0.8" />
     <PackageReference Include="Microsoft.ApplicationInsights" Version="2.23.0" />
-	<!-- Upgrading to 5.3.5 will cause the function to fail, must upgrade function from .Net 6 first-->
-    <PackageReference Include="Microsoft.Azure.WebJobs.Extensions.Storage.Queues" Version="5.3.5" />
-    <PackageReference Include="Microsoft.NET.Sdk.Functions" Version="4.6.0" />
   </ItemGroup>
 
   <ItemGroup Condition="'$(Configuration)'=='Debug'">
@@ -28,20 +25,6 @@
     <AdditionalFiles Include="..\..\stylecop.json">
       <Link>stylecop.json</Link>
     </AdditionalFiles>
-  </ItemGroup>
-
-  <PropertyGroup>
-    <_FunctionsSkipCleanOutput>true</_FunctionsSkipCleanOutput>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <None Update="host.json">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </None>
-    <None Update="local.settings.json">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-      <CopyToPublishDirectory>Never</CopyToPublishDirectory>
-    </None>
   </ItemGroup>
 
   <ItemGroup>
