+++ src/Events.Functions/Services/CertificateResolverService.cs
@@ -1,5 +1,4 @@
 using System;
-using System.Diagnostics.CodeAnalysis;
 using System.Security.Cryptography.X509Certificates;
 using System.Threading.Tasks;
 using Altinn.Platform.Events.Functions.Configuration;
